Create Multiome_clusterandmerge
The first step to analyzing multiome data is to construct a seurat/signac object containing the RNA and ATAC data. This code demonstrates how to do that for 4 datasets. 

Merging was functional for this dataset due to low batch effect. Resulting UMAPs should be examined for batch effects prior to continued analysis.